requirements.txt
streamlit
numpy
opencv-python
easyocr
Pillow
pywhatkit
